BackEnd/requirements.txt
django==3.2.4
requests==2.25.1
psycopg2==2.8.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
